+++ src/actions.ts
@@ -1,43 +1,52 @@
 import puppeteer from 'puppeteer'
 import Infrastructure from './infrastructure'
+import { prepareStackTrace } from './utils/error'
 
 export default class Actions extends Infrastructure {
   click (selector: string) {
-    this.push(async () => await this.page.click(selector))
+    this.push(async () => await this.page.click(selector), prepareStackTrace())
 
     return this
   }
 
   doubleClick (selector: string) {
-    this.push(async () => {
-      await this.page.click(selector, { clickCount: 2 })
-    })
+    this.push(
+      async () => await this.page.click(selector, { clickCount: 2 }),
+      prepareStackTrace()
+    )
 
     return this
   }
 
   rightClick (selector: string) {
-    this.push(async () => {
-      await this.page.click(selector, { button: 'right' })
-    })
+    this.push(
+      async () => await this.page.click(selector, { button: 'right' }),
+      prepareStackTrace()
+    )
 
     return this
   }
 
   hover (selector: string) {
-    this.push(async () => await this.page.hover(selector))
+    this.push(async () => await this.page.hover(selector), prepareStackTrace())
 
     return this
   }
 
   type (selector: string, text: string) {
-    this.push(async () => await this.page.type(selector, text))
+    this.push(
+      async () => await this.page.type(selector, text),
+      prepareStackTrace()
+    )
 
     return this
   }
 
   sendChar (char: string) {
-    this.push(async () => await this.page.keyboard.sendCharacter(char))
+    this.push(
+      async () => await this.page.keyboard.sendCharacter(char),
+      prepareStackTrace()
+    )
 
     return this
   }
@@ -61,13 +70,13 @@ export default class Actions extends Infrastructure {
         },
         selector
       )
-    })
+    }, prepareStackTrace())
 
     return this
   }
 
   focus (selector: string) {
-    this.push(async () => await this.page.focus(selector))
+    this.push(async () => await this.page.focus(selector), prepareStackTrace())
 
     return this
   }
@@ -79,7 +88,7 @@ export default class Actions extends Infrastructure {
       } else {
         await this.page.select(selector, values)
       }
-    })
+    }, prepareStackTrace())
 
     return this
   }
@@ -91,7 +100,7 @@ export default class Actions extends Infrastructure {
         sel => document.querySelector<HTMLInputElement>(sel)!.checked = true,
         selector
       )
-    })
+    }, prepareStackTrace())
 
     return this
   }
@@ -103,7 +112,7 @@ export default class Actions extends Infrastructure {
         sel => document.querySelector<HTMLInputElement>(sel)!.checked = false,
         selector
       )
-    })
+    }, prepareStackTrace())
 
     return this
   }
@@ -118,7 +127,7 @@ export default class Actions extends Infrastructure {
         selector,
         value
       )
-    })
+    }, prepareStackTrace())
 
     return this
   }
@@ -130,49 +139,64 @@ export default class Actions extends Infrastructure {
       } else {
         await this.page.keyboard.press(key)
       }
-    })
+    }, prepareStackTrace())
 
     return this
   }
 
   keyDown (key: string) {
-    this.push(async () => await this.page.keyboard.down(key))
+    this.push(
+      async () => await this.page.keyboard.down(key),
+      prepareStackTrace()
+    )
 
     return this
   }
 
   keyUp (key: string) {
-    this.push(async () => await this.page.keyboard.up(key))
+    this.push(async () => await this.page.keyboard.up(key), prepareStackTrace())
 
     return this
   }
 
   mouseMoveTo (x: number, y: number) {
-    this.push(async () => await this.page.mouse.move(x, y))
+    this.push(async () => await this.page.mouse.move(x, y), prepareStackTrace())
 
     return this
   }
 
   mouseClick (x: number, y: number, options?: puppeteer.MousePressOptions) {
-    this.push(async () => await this.page.mouse.click(x, y, options))
+    this.push(
+      async () => await this.page.mouse.click(x, y, options),
+      prepareStackTrace()
+    )
 
     return this
   }
 
   mouseDown (button: puppeteer.MouseButtons = 'left', clickCount: number = 1) {
-    this.push(async () => await this.page.mouse.down({ button, clickCount }))
+    this.push(
+      async () => await this.page.mouse.down({ button, clickCount }),
+      prepareStackTrace()
+    )
 
     return this
   }
 
   mouseUp (button: puppeteer.MouseButtons = 'left', clickCount: number = 1) {
-    this.push(async () => await this.page.mouse.up({ button, clickCount }))
+    this.push(
+      async () => await this.page.mouse.up({ button, clickCount }),
+      prepareStackTrace()
+    )
 
     return this
   }
 
   uploadFile (selector: string, path: string) {
-    this.push(async () => await (await this.page.$(selector))!.uploadFile(path))
+    this.push(
+      async () => await (await this.page.$(selector))!.uploadFile(path),
+      prepareStackTrace()
+    )
 
     return this
   }
@@ -186,7 +210,7 @@ export default class Actions extends Infrastructure {
         selector,
         className
       )
-    })
+    }, prepareStackTrace())
 
     return this
   }
@@ -200,7 +224,7 @@ export default class Actions extends Infrastructure {
         selector,
         className
       )
-    })
+    }, prepareStackTrace())
 
     return this
   }
@@ -214,19 +238,25 @@ export default class Actions extends Infrastructure {
         selector,
         className
       )
-    })
+    }, prepareStackTrace())
 
     return this
   }
 
   setCookie (...cookies: puppeteer.SetCookie[]) {
-    this.push(async () => await this.page.setCookie(...cookies))
+    this.push(
+      async () => await this.page.setCookie(...cookies),
+      prepareStackTrace()
+    )
 
     return this
   }
 
   deleteCookie (...cookies: puppeteer.DeleteCookie[]) {
-    this.push(async () => await this.page.deleteCookie(...cookies))
+    this.push(
+      async () => await this.page.deleteCookie(...cookies),
+      prepareStackTrace()
+    )
 
     return this
   }