+++ src/utils/error.ts
@@ -1,0 +1,16 @@
+export function prepareStackTrace () {
+  return new Error()
+}
+
+/**
+ * @param fake The error which collects stack information
+ * @param real The real error which contains error message.
+ */
+export function beautifyStack (fake: Error, real: Error) {
+  const stack = fake.stack!.split('\n')
+  stack.splice(1, 2)
+  stack[0] = real.message
+  real.stack = stack.join('\n')
+
+  return real
+}