+++ tests/utils/error.test.ts
@@ -1,0 +1,14 @@
+import * as error from '../../src/utils/error'
+
+test('beautify error stack', () => {
+  const fake = new Error()
+  const real = new Error()
+
+  fake.stack = [1, 2, 3, 4, 5].join('\n')
+  real.message = 'real'
+
+  const beautified = error.beautifyStack(fake, real)
+  expect(beautified).toBe(real)
+  expect(real.message).toBe('real')
+  expect(real.stack).toBe('real\n4\n5')
+})